Rewrote talking to server, added some functions? that appeared in new server

+++ app/src/main/java/com/example/asus/taskmanager/TaskShowFragment.java
@@ -4,10 +4,6 @@ import android.app.Activity;
 import android.app.DatePickerDialog;
 import android.app.Fragment;
 import android.app.TimePickerDialog;
-import android.content.Context;
-import android.content.Intent;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.view.LayoutInflater;
@@ -19,10 +15,7 @@ import android.widget.TextView;
 import android.widget.TimePicker;
 import android.widget.Toast;
 
-import java.text.DateFormat;
 import java.util.Calendar;
-import java.util.Date;
-import java.util.Locale;
 
 public class TaskShowFragment extends Fragment  {
 
@@ -131,7 +124,7 @@ public class TaskShowFragment extends Fragment  {
         return view;
     }
 
-    public void setIndex(int index) {
+    public void setIndex(long index) {
         this.index = index;
     }
